requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
anthropic>=0.3.0
pytube>=12.1.0
youtube-transcript-api>=0.6.0
selenium>=4.10.0
Pillow>=9.5.0
requests>=2.31.0
opencv-python>=4.8.0
numpy>=1.24.0
python-multipart>=0.0.6
aiofiles>=0.8.0
yt-dlp>=2023.0.0
tesseract
pytesseract
playwright
google-api-python-client
moviepy>=1.0.3
#brew install tesseract